Size now included in item name and slug
If an item has a size, it is now paranthetically appended to the item's name. It is also used to prepoulate the slug field. This makes it easier to add items of the same make and model but different sizes into the system, without needing to include the size in both the item's size field and the model field.